+++ breezeml/__init__.py
@@ -2,7 +2,7 @@
 🌬️ BreezeML 🔥✨
 Beginner-friendly machine learning library built on scikit-learn.
 """
-__version__ = "0.1.2"
+__version__ = "0.2.0"
 __author__ = "Akash Anipakalu Giridhar 🔥✨"
 
 from .breezeml import fit, predict, auto, save, load, datasets, creator, classify, regress, from_csv, report